Split daemon building into seperate module
The main module should only be daemon code, not also builder code.
